fix: Moved security files into security folder and removed unused import
